Create a new feature
No idea how to add a new request to your screenshot tool, so i'll just write it here sorry about this. 
I like the tool that it's very simple and doesn't have too much garbage in it.
I would like in Area Capture, the checkmark and X icon is good but how about add a 2 more icons a little pen and a little blur.
With mouse it would be possible to circle something with pen or draw something or highlight something and
you could remove some sensitive information with blur.